requirements.txt
Box2D==2.3.2
box2d-py==2.3.5
cloudpickle==3.0.0
Farama-Notifications==0.0.4
gymnasium==0.29.1
numpy==1.26.3
pillow==10.2.0
pygame==2.5.2
setuptools==69.1.0
tk==0.1.0
swig==4.1.1.post1
typing_extensions==4.9.0
wheel==0.42.0 